@@ -1,85 +1,47 @@
 {
-  "_from": "accepts@~1.3.4",
-  "_id": "accepts@1.3.4",
-  "_inBundle": false,
-  "_integrity": "sha1-hiRnWMfdbSGmR0/whKR0DsBesh8=",
-  "_location": "/accepts",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "accepts@~1.3.4",
-    "name": "accepts",
-    "escapedName": "accepts",
-    "rawSpec": "~1.3.4",
-    "saveSpec": null,
-    "fetchSpec": "~1.3.4"
-  },
-  "_requiredBy": [
-    "/express"
-  ],
-  "_resolved": "https://registry.npmjs.org/accepts/-/accepts-1.3.4.tgz",
-  "_shasum": "86246758c7dd6d21a6474ff084a4740ec05eb21f",
-  "_spec": "accepts@~1.3.4",
-  "_where": "/Users/harikiranyalavarthi/Dropbox/Coursera/basic-auth/node_modules/express",
-  "bugs": {
-    "url": "https://github.com/jshttp/accepts/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "Jonathan Ong",
-      "email": "me@jongleberry.com",
-      "url": "http://jongleberry.com"
-    }
-  ],
-  "dependencies": {
-    "mime-types": "~2.1.16",
-    "negotiator": "0.6.1"
-  },
-  "deprecated": false,
+  "name": "accepts",
   "description": "Higher-level content negotiation",
-  "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
-    "istanbul": "0.4.5",
-    "mocha": "~1.21.5"
+  "version": "1.3.8",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
+  ],
+  "license": "MIT",
+  "repository": "jshttp/accepts",
+  "dependencies": {
+    "mime-types": "~2.1.34",
+    "negotiator": "0.6.3"
   },
-  "engines": {
-    "node": ">= 0.6"
+  "devDependencies": {
+    "deep-equal": "1.0.1",
+    "eslint": "7.32.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.25.4",
+    "eslint-plugin-markdown": "2.2.1",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "4.3.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "9.2.0",
+    "nyc": "15.1.0"
   },
   "files": [
     "LICENSE",
     "HISTORY.md",
     "index.js"
   ],
-  "homepage": "https://github.com/jshttp/accepts#readme",
+  "engines": {
+    "node": ">= 0.6"
+  },
+  "scripts": {
+    "lint": "eslint .",
+    "test": "mocha --reporter spec --check-leaks --bail test/",
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
+  },
   "keywords": [
     "content",
     "negotiation",
     "accept",
     "accepts"
-  ],
-  "license": "MIT",
-  "name": "accepts",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/accepts.git"
-  },
-  "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
-    "test": "mocha --reporter spec --check-leaks --bail test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
-  },
-  "version": "1.3.4"
+  ]
 }
