@@ -1,73 +1,41 @@
 {
-  "_from": "basic-auth@~2.0.0",
-  "_id": "basic-auth@2.0.0",
-  "_inBundle": false,
-  "_integrity": "sha1-AV2z81PgLlY3d1X5YnQuiYHnu7o=",
-  "_location": "/basic-auth",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "basic-auth@~2.0.0",
-    "name": "basic-auth",
-    "escapedName": "basic-auth",
-    "rawSpec": "~2.0.0",
-    "saveSpec": null,
-    "fetchSpec": "~2.0.0"
-  },
-  "_requiredBy": [
-    "/morgan"
-  ],
-  "_resolved": "https://registry.npmjs.org/basic-auth/-/basic-auth-2.0.0.tgz",
-  "_shasum": "015db3f353e02e56377755f962742e8981e7bbba",
-  "_spec": "basic-auth@~2.0.0",
-  "_where": "/Users/harikiranyalavarthi/Dropbox/Coursera/basic-auth/node_modules/morgan",
-  "bugs": {
-    "url": "https://github.com/jshttp/basic-auth/issues"
-  },
-  "bundleDependencies": false,
-  "dependencies": {
-    "safe-buffer": "5.1.1"
-  },
-  "deprecated": false,
+  "name": "basic-auth",
   "description": "node.js basic auth parser",
-  "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
-    "istanbul": "0.4.5",
-    "mocha": "2.5.3"
-  },
-  "engines": {
-    "node": ">= 0.8"
-  },
-  "files": [
-    "HISTORY.md",
-    "LICENSE",
-    "index.js"
-  ],
-  "homepage": "https://github.com/jshttp/basic-auth#readme",
+  "version": "2.0.1",
+  "license": "MIT",
   "keywords": [
     "basic",
     "auth",
     "authorization",
     "basicauth"
   ],
-  "license": "MIT",
-  "name": "basic-auth",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/basic-auth.git"
+  "repository": "jshttp/basic-auth",
+  "dependencies": {
+    "safe-buffer": "5.1.2"
+  },
+  "devDependencies": {
+    "eslint": "5.6.0",
+    "eslint-config-standard": "12.0.0",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-markdown": "1.0.0-beta.6",
+    "eslint-plugin-node": "7.0.1",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
+    "istanbul": "0.4.5",
+    "mocha": "5.2.0"
+  },
+  "files": [
+    "HISTORY.md",
+    "LICENSE",
+    "index.js"
+  ],
+  "engines": {
+    "node": ">= 0.8"
   },
   "scripts": {
     "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --check-leaks --reporter spec --bail",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
-  },
-  "version": "2.0.0"
+  }
 }
