@@ -1,91 +1,56 @@
 {
-  "_from": "body-parser@1.18.2",
-  "_id": "body-parser@1.18.2",
-  "_inBundle": false,
-  "_integrity": "sha1-h2eKGdhLR9hZuDGZvVm84iKxBFQ=",
-  "_location": "/body-parser",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "version",
-    "registry": true,
-    "raw": "body-parser@1.18.2",
-    "name": "body-parser",
-    "escapedName": "body-parser",
-    "rawSpec": "1.18.2",
-    "saveSpec": null,
-    "fetchSpec": "1.18.2"
-  },
-  "_requiredBy": [
-    "/express"
-  ],
-  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.18.2.tgz",
-  "_shasum": "87678a19d84b47d859b83199bd59bce222b10454",
-  "_spec": "body-parser@1.18.2",
-  "_where": "/Users/harikiranyalavarthi/Dropbox/Coursera/basic-auth/node_modules/express",
-  "bugs": {
-    "url": "https://github.com/expressjs/body-parser/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "Jonathan Ong",
-      "email": "me@jongleberry.com",
-      "url": "http://jongleberry.com"
-    }
-  ],
-  "dependencies": {
-    "bytes": "3.0.0",
-    "content-type": "~1.0.4",
-    "debug": "2.6.9",
-    "depd": "~1.1.1",
-    "http-errors": "~1.6.2",
-    "iconv-lite": "0.4.19",
-    "on-finished": "~2.3.0",
-    "qs": "6.5.1",
-    "raw-body": "2.3.2",
-    "type-is": "~1.6.15"
-  },
-  "deprecated": false,
+  "name": "body-parser",
   "description": "Node.js body parsing middleware",
-  "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
-    "istanbul": "0.4.5",
-    "methods": "1.1.2",
-    "mocha": "2.5.3",
-    "safe-buffer": "5.1.1",
-    "supertest": "1.1.0"
+  "version": "1.20.2",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
+  ],
+  "license": "MIT",
+  "repository": "expressjs/body-parser",
+  "dependencies": {
+    "bytes": "3.1.2",
+    "content-type": "~1.0.5",
+    "debug": "2.6.9",
+    "depd": "2.0.0",
+    "destroy": "1.2.0",
+    "http-errors": "2.0.0",
+    "iconv-lite": "0.4.24",
+    "on-finished": "2.4.1",
+    "qs": "6.11.0",
+    "raw-body": "2.5.2",
+    "type-is": "~1.6.18",
+    "unpipe": "1.0.0"
   },
-  "engines": {
-    "node": ">= 0.8"
+  "devDependencies": {
+    "eslint": "8.34.0",
+    "eslint-config-standard": "14.1.1",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-markdown": "3.0.0",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-standard": "4.1.0",
+    "methods": "1.1.2",
+    "mocha": "10.2.0",
+    "nyc": "15.1.0",
+    "safe-buffer": "5.2.1",
+    "supertest": "6.3.3"
   },
   "files": [
     "lib/",
     "LICENSE",
     "HISTORY.md",
+    "SECURITY.md",
     "index.js"
   ],
-  "homepage": "https://github.com/expressjs/body-parser#readme",
-  "license": "MIT",
-  "name": "body-parser",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/expressjs/body-parser.git"
+  "engines": {
+    "node": ">= 0.8",
+    "npm": "1.2.8000 || >= 1.4.16"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --require test/support/env --reporter spec --check-leaks --bail test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --require test/support/env --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --require test/support/env --reporter spec --check-leaks test/"
-  },
-  "version": "1.18.2"
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
+  }
 }
