@@ -1,58 +1,9 @@
 {
-  "_from": "content-type@~1.0.4",
-  "_id": "content-type@1.0.4",
-  "_inBundle": false,
-  "_integrity": "sha512-hIP3EEPs8tB9AT1L+NUqtwOAps4mk2Zob89MWXMHjHWg9milF/j4osnnQLXBCBFBk/tvIG/tUc9mOUJiPBhPXA==",
-  "_location": "/content-type",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "content-type@~1.0.4",
-    "name": "content-type",
-    "escapedName": "content-type",
-    "rawSpec": "~1.0.4",
-    "saveSpec": null,
-    "fetchSpec": "~1.0.4"
-  },
-  "_requiredBy": [
-    "/body-parser",
-    "/express"
-  ],
-  "_resolved": "https://registry.npmjs.org/content-type/-/content-type-1.0.4.tgz",
-  "_shasum": "e138cc75e040c727b1966fe5e5f8c9aee256fe3b",
-  "_spec": "content-type@~1.0.4",
-  "_where": "/Users/harikiranyalavarthi/Dropbox/Coursera/basic-auth/node_modules/express",
-  "author": {
-    "name": "Douglas Christopher Wilson",
-    "email": "doug@somethingdoug.com"
-  },
-  "bugs": {
-    "url": "https://github.com/jshttp/content-type/issues"
-  },
-  "bundleDependencies": false,
-  "deprecated": false,
+  "name": "content-type",
   "description": "Create and parse HTTP Content-Type header",
-  "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
-    "istanbul": "0.4.5",
-    "mocha": "~1.21.5"
-  },
-  "engines": {
-    "node": ">= 0.6"
-  },
-  "files": [
-    "LICENSE",
-    "HISTORY.md",
-    "README.md",
-    "index.js"
-  ],
-  "homepage": "https://github.com/jshttp/content-type#readme",
+  "version": "1.0.5",
+  "author": "Douglas Christopher Wilson <doug@somethingdoug.com>",
+  "license": "MIT",
   "keywords": [
     "content-type",
     "http",
@@ -60,17 +11,32 @@
     "res",
     "rfc7231"
   ],
-  "license": "MIT",
-  "name": "content-type",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jshttp/content-type.git"
+  "repository": "jshttp/content-type",
+  "devDependencies": {
+    "deep-equal": "1.0.1",
+    "eslint": "8.32.0",
+    "eslint-config-standard": "15.0.1",
+    "eslint-plugin-import": "2.27.5",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-standard": "4.1.0",
+    "mocha": "10.2.0",
+    "nyc": "15.1.0"
+  },
+  "files": [
+    "LICENSE",
+    "HISTORY.md",
+    "README.md",
+    "index.js"
+  ],
+  "engines": {
+    "node": ">= 0.6"
   },
   "scripts": {
     "lint": "eslint .",
     "test": "mocha --reporter spec --check-leaks --bail test/",
-    "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/"
-  },
-  "version": "1.0.4"
+    "test-ci": "nyc --reporter=lcovonly --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test",
+    "version": "node scripts/version-history.js && git add HISTORY.md"
+  }
 }
