@@ -1,52 +1,26 @@
 {
-  "_from": "encodeurl@~1.0.1",
-  "_id": "encodeurl@1.0.1",
-  "_inBundle": false,
-  "_integrity": "sha1-eePVhlU0aQn+bw9Fpd5oEDspTSA=",
-  "_location": "/encodeurl",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "range",
-    "registry": true,
-    "raw": "encodeurl@~1.0.1",
-    "name": "encodeurl",
-    "escapedName": "encodeurl",
-    "rawSpec": "~1.0.1",
-    "saveSpec": null,
-    "fetchSpec": "~1.0.1"
-  },
-  "_requiredBy": [
-    "/express",
-    "/finalhandler",
-    "/send",
-    "/serve-static"
-  ],
-  "_resolved": "https://registry.npmjs.org/encodeurl/-/encodeurl-1.0.1.tgz",
-  "_shasum": "79e3d58655346909fe6f0f45a5de68103b294d20",
-  "_spec": "encodeurl@~1.0.1",
-  "_where": "/Users/harikiranyalavarthi/Dropbox/Coursera/basic-auth/node_modules/express",
-  "bugs": {
-    "url": "https://github.com/pillarjs/encodeurl/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    }
-  ],
-  "deprecated": false,
+  "name": "encodeurl",
   "description": "Encode a URL to a percent-encoded form, excluding already-encoded sequences",
+  "version": "1.0.2",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>"
+  ],
+  "license": "MIT",
+  "keywords": [
+    "encode",
+    "encodeurl",
+    "url"
+  ],
+  "repository": "pillarjs/encodeurl",
   "devDependencies": {
-    "eslint": "2.11.1",
-    "eslint-config-standard": "5.3.1",
-    "eslint-plugin-promise": "1.3.2",
-    "eslint-plugin-standard": "1.3.2",
-    "istanbul": "0.4.3",
+    "eslint": "3.19.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-node": "5.2.1",
+    "eslint-plugin-promise": "3.6.0",
+    "eslint-plugin-standard": "3.0.1",
+    "istanbul": "0.4.5",
     "mocha": "2.5.3"
-  },
-  "engines": {
-    "node": ">= 0.8"
   },
   "files": [
     "LICENSE",
@@ -54,23 +28,13 @@
     "README.md",
     "index.js"
   ],
-  "homepage": "https://github.com/pillarjs/encodeurl#readme",
-  "keywords": [
-    "encode",
-    "encodeurl",
-    "url"
-  ],
-  "license": "MIT",
-  "name": "encodeurl",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/pillarjs/encodeurl.git"
+  "engines": {
+    "node": ">= 0.8"
   },
   "scripts": {
-    "lint": "eslint **/*.js",
+    "lint": "eslint .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
-  },
-  "version": "1.0.1"
+  }
 }
