@@ -1,91 +1,47 @@
 {
-  "_from": "express-session",
-  "_id": "express-session@1.15.6",
-  "_inBundle": false,
-  "_integrity": "sha512-r0nrHTCYtAMrFwZ0kBzZEXa1vtPVrw0dKvGSrKP4dahwBQ1BJpF2/y1Pp4sCD/0kvxV4zZeclyvfmw0B4RMJQA==",
-  "_location": "/express-session",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "tag",
-    "registry": true,
-    "raw": "express-session",
-    "name": "express-session",
-    "escapedName": "express-session",
-    "rawSpec": "",
-    "saveSpec": null,
-    "fetchSpec": "latest"
-  },
-  "_requiredBy": [
-    "#USER",
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/express-session/-/express-session-1.15.6.tgz",
-  "_shasum": "47b4160c88f42ab70fe8a508e31cbff76757ab0a",
-  "_spec": "express-session",
-  "_where": "/Users/harikiranyalavarthi/Dropbox/Coursera/basic-auth",
-  "author": {
-    "name": "TJ Holowaychuk",
-    "email": "tj@vision-media.ca",
-    "url": "http://tjholowaychuk.com"
-  },
-  "bugs": {
-    "url": "https://github.com/expressjs/session/issues"
-  },
-  "bundleDependencies": false,
-  "contributors": [
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "Joe Wagner",
-      "email": "njwjs722@gmail.com"
-    }
-  ],
-  "dependencies": {
-    "cookie": "0.3.1",
-    "cookie-signature": "1.0.6",
-    "crc": "3.4.4",
-    "debug": "2.6.9",
-    "depd": "~1.1.1",
-    "on-headers": "~1.0.1",
-    "parseurl": "~1.3.2",
-    "uid-safe": "~2.1.5",
-    "utils-merge": "1.0.1"
-  },
-  "deprecated": false,
+  "name": "express-session",
+  "version": "1.18.0",
   "description": "Simple session middleware for Express",
+  "author": "TJ Holowaychuk <tj@vision-media.ca> (http://tjholowaychuk.com)",
+  "contributors": [
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "Joe Wagner <njwjs722@gmail.com>"
+  ],
+  "repository": "expressjs/session",
+  "license": "MIT",
+  "dependencies": {
+    "cookie": "0.6.0",
+    "cookie-signature": "1.0.7",
+    "debug": "2.6.9",
+    "depd": "~2.0.0",
+    "on-headers": "~1.0.2",
+    "parseurl": "~1.3.3",
+    "safe-buffer": "5.2.1",
+    "uid-safe": "~2.1.5"
+  },
   "devDependencies": {
     "after": "0.8.2",
-    "cookie-parser": "1.4.3",
-    "eslint": "3.19.0",
-    "eslint-plugin-markdown": "1.0.0-beta.6",
-    "express": "4.15.5",
-    "istanbul": "0.4.5",
-    "mocha": "2.5.3",
-    "supertest": "1.1.0"
-  },
-  "engines": {
-    "node": ">= 0.8.0"
+    "cookie-parser": "1.4.6",
+    "eslint": "8.56.0",
+    "eslint-plugin-markdown": "3.0.1",
+    "express": "4.17.3",
+    "mocha": "10.2.0",
+    "nyc": "15.1.0",
+    "supertest": "6.3.4"
   },
   "files": [
     "session/",
     "HISTORY.md",
-    "LICENSE",
     "index.js"
   ],
-  "homepage": "https://github.com/expressjs/session#readme",
-  "license": "MIT",
-  "name": "express-session",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/expressjs/session.git"
+  "engines": {
+    "node": ">= 0.8.0"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
-    "test": "mocha --check-leaks --bail --no-exit --reporter spec test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --check-leaks --no-exit --reporter dot test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --check-leaks --no-exit --reporter spec test/"
-  },
-  "version": "1.15.6"
+    "lint": "eslint . && node ./scripts/lint-readme.js",
+    "test": "mocha --require test/support/env --check-leaks --bail --no-exit --reporter spec test/",
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc npm test",
+    "version": "node scripts/version-history.js && git add HISTORY.md"
+  }
 }
