@@ -1,118 +1,81 @@
 {
-  "_from": "express",
-  "_id": "express@4.16.2",
-  "_inBundle": false,
-  "_integrity": "sha1-41xt/i1kt9ygpc1PIXgb4ymeB2w=",
-  "_location": "/express",
-  "_phantomChildren": {},
-  "_requested": {
-    "type": "tag",
-    "registry": true,
-    "raw": "express",
-    "name": "express",
-    "escapedName": "express",
-    "rawSpec": "",
-    "saveSpec": null,
-    "fetchSpec": "latest"
-  },
-  "_requiredBy": [
-    "#USER",
-    "/"
-  ],
-  "_resolved": "https://registry.npmjs.org/express/-/express-4.16.2.tgz",
-  "_shasum": "e35c6dfe2d64b7dca0a5cd4f21781be3299e076c",
-  "_spec": "express",
-  "_where": "/Users/harikiranyalavarthi/Dropbox/Coursera/basic-auth",
-  "author": {
-    "name": "TJ Holowaychuk",
-    "email": "tj@vision-media.ca"
-  },
-  "bugs": {
-    "url": "https://github.com/expressjs/express/issues"
-  },
-  "bundleDependencies": false,
+  "name": "express",
+  "description": "Fast, unopinionated, minimalist web framework",
+  "version": "4.19.2",
+  "author": "TJ Holowaychuk <tj@vision-media.ca>",
   "contributors": [
-    {
-      "name": "Aaron Heckmann",
-      "email": "aaron.heckmann+github@gmail.com"
-    },
-    {
-      "name": "Ciaran Jessup",
-      "email": "ciaranj@gmail.com"
-    },
-    {
-      "name": "Douglas Christopher Wilson",
-      "email": "doug@somethingdoug.com"
-    },
-    {
-      "name": "Guillermo Rauch",
-      "email": "rauchg@gmail.com"
-    },
-    {
-      "name": "Jonathan Ong",
-      "email": "me@jongleberry.com"
-    },
-    {
-      "name": "Roman Shtylman",
-      "email": "shtylman+expressjs@gmail.com"
-    },
-    {
-      "name": "Young Jae Sim",
-      "email": "hanul@hanul.me"
-    }
+    "Aaron Heckmann <aaron.heckmann+github@gmail.com>",
+    "Ciaran Jessup <ciaranj@gmail.com>",
+    "Douglas Christopher Wilson <doug@somethingdoug.com>",
+    "Guillermo Rauch <rauchg@gmail.com>",
+    "Jonathan Ong <me@jongleberry.com>",
+    "Roman Shtylman <shtylman+expressjs@gmail.com>",
+    "Young Jae Sim <hanul@hanul.me>"
+  ],
+  "license": "MIT",
+  "repository": "expressjs/express",
+  "homepage": "http://expressjs.com/",
+  "keywords": [
+    "express",
+    "framework",
+    "sinatra",
+    "web",
+    "http",
+    "rest",
+    "restful",
+    "router",
+    "app",
+    "api"
   ],
   "dependencies": {
-    "accepts": "~1.3.4",
+    "accepts": "~1.3.8",
     "array-flatten": "1.1.1",
-    "body-parser": "1.18.2",
-    "content-disposition": "0.5.2",
+    "body-parser": "1.20.2",
+    "content-disposition": "0.5.4",
     "content-type": "~1.0.4",
-    "cookie": "0.3.1",
+    "cookie": "0.6.0",
     "cookie-signature": "1.0.6",
     "debug": "2.6.9",
-    "depd": "~1.1.1",
-    "encodeurl": "~1.0.1",
+    "depd": "2.0.0",
+    "encodeurl": "~1.0.2",
     "escape-html": "~1.0.3",
     "etag": "~1.8.1",
-    "finalhandler": "1.1.0",
+    "finalhandler": "1.2.0",
     "fresh": "0.5.2",
+    "http-errors": "2.0.0",
     "merge-descriptors": "1.0.1",
     "methods": "~1.1.2",
-    "on-finished": "~2.3.0",
-    "parseurl": "~1.3.2",
+    "on-finished": "2.4.1",
+    "parseurl": "~1.3.3",
     "path-to-regexp": "0.1.7",
-    "proxy-addr": "~2.0.2",
-    "qs": "6.5.1",
-    "range-parser": "~1.2.0",
-    "safe-buffer": "5.1.1",
-    "send": "0.16.1",
-    "serve-static": "1.13.1",
-    "setprototypeof": "1.1.0",
-    "statuses": "~1.3.1",
-    "type-is": "~1.6.15",
+    "proxy-addr": "~2.0.7",
+    "qs": "6.11.0",
+    "range-parser": "~1.2.1",
+    "safe-buffer": "5.2.1",
+    "send": "0.18.0",
+    "serve-static": "1.15.0",
+    "setprototypeof": "1.2.0",
+    "statuses": "2.0.1",
+    "type-is": "~1.6.18",
     "utils-merge": "1.0.1",
     "vary": "~1.1.2"
   },
-  "deprecated": false,
-  "description": "Fast, unopinionated, minimalist web framework",
   "devDependencies": {
     "after": "0.8.2",
-    "connect-redis": "~2.4.1",
-    "cookie-parser": "~1.4.3",
-    "cookie-session": "1.3.2",
-    "ejs": "2.5.7",
-    "eslint": "2.13.1",
-    "express-session": "1.15.6",
-    "hbs": "4.0.1",
-    "istanbul": "0.4.5",
-    "marked": "0.3.6",
-    "method-override": "2.3.10",
-    "mocha": "3.5.3",
-    "morgan": "1.9.0",
-    "multiparty": "4.1.3",
+    "connect-redis": "3.4.2",
+    "cookie-parser": "1.4.6",
+    "cookie-session": "2.0.0",
+    "ejs": "3.1.9",
+    "eslint": "8.47.0",
+    "express-session": "1.17.2",
+    "hbs": "4.2.0",
+    "marked": "0.7.0",
+    "method-override": "3.0.0",
+    "mocha": "10.2.0",
+    "morgan": "1.10.0",
+    "nyc": "15.1.0",
     "pbkdf2-password": "1.2.1",
-    "should": "13.1.0",
-    "supertest": "1.2.0",
+    "supertest": "6.3.0",
     "vhost": "~3.0.2"
   },
   "engines": {
