@@ -1,14 +1,20 @@
-var express = require('express');
-var morgan = require('morgan');
-var session = require('express-session');
-var FileStore = require('session-file-store')(session);
+import express from 'express';
+import morgan from 'morgan';
+import session from 'express-session';
+import sessionFileStore from 'session-file-store';
+import { fileURLToPath } from 'url';
+import { dirname } from 'path';
 
-var hostname = 'localhost';
-var port = 3000;
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
-var app = express();
+const FileStore = sessionFileStore(session);
+const hostname = 'localhost';
+const port = 3001;
+const app = express();
 
 app.use(morgan('dev'));
+
 app.use(session({
   name: 'session-id',
   secret: '12345-67890-09876-54321',
@@ -17,24 +23,24 @@ app.use(session({
   store: new FileStore()
 }));
 
-function auth (req, res, next) {
+const auth = (req, res, next) => {
     console.log(req.headers);
     if (!req.session.user) {
-        var authHeader = req.headers.authorization;
+        const authHeader = req.headers.authorization;
         if (!authHeader) {
-            var err = new Error('You are not authenticated!');
+            const err = new Error('You are not authenticated!');
             err.status = 401;
             next(err);
             return;
         }
-        var auth = new Buffer(authHeader.split(' ')[1], 'base64').toString().split(':');
-        var user = auth[0];
-        var pass = auth[1];
-        if (user == 'admin' && pass == 'password') {
+        const auth = Buffer.from(authHeader.split(' ')[1], 'base64').toString().split(':');
+        const user = auth[0];
+        const pass = auth[1];
+        if (user === 'admin' && pass === 'password') {
             req.session.user = 'admin';
             next(); // authorized
         } else {
-            var err = new Error('You are not authenticated!');
+            const err = new Error('You are not authenticated!');
             err.status = 401;
             next(err);
         }
@@ -45,24 +51,25 @@ function auth (req, res, next) {
             next();
         }
         else {
-            var err = new Error('You are not authenticated!');
+            const err = new Error('You are not authenticated!');
             err.status = 401;
             next(err);
         }
-    }}
+    }
+};
 
 app.use(auth);
 
-app.use(express.static(__dirname + '/public'));
+app.use(express.static(`${__dirname}/public`));
 
-app.use(function(err,req,res,next) {
-            res.writeHead(err.status || 500, {
-            'WWW-Authenticate': 'Basic',
-            'Content-Type': 'text/plain'
-        });
-        res.end(err.message);
+app.use((err, req, res, next) => {
+    res.writeHead(err.status || 500, {
+        'WWW-Authenticate': 'Basic',
+        'Content-Type': 'text/plain'
+    });
+    res.end(err.message);
 });
 
-app.listen(port, hostname, function(){
+app.listen(port, hostname, () => {
   console.log(`Server running at http://${hostname}:${port}/`);
 });
